imitation-in-homes/requirements.txt
byol_pytorch==0.5.7
einops==0.7.0
hydra-core==1.3.2
matplotlib==3.5.1
numpy_quaternion==2022.4.2
omegaconf==2.3.0
Pillow==9.5.0
pyliblzfse==0.4.1
scipy==1.10.0
tqdm==4.65.0
accelerate
timm==0.9.2
wandb==0.15.8
vit_pytorch==1.6.4